refactor: Move location translation to its own module
